requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
distro==1.9.0
entrypoints==0.4
faiss-cpu==1.7.4
Faker==25.2.0
favicon==0.7.0
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.57
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
openai==1.30.1
openapi-schema-pydantic==1.2.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
prometheus_client==0.20.0
protobuf==3.20.3
pyarrow==16.1.0
pydantic==1.10.15
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.8.1
Pympler==1.0.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
streamlit==1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
yarl==1.9.4
zipp==3.18.1
